# .internal/docs/swagger.yaml
basePath: /
definitions:
  auth.Account:
    properties:
      roles:
        items:
          type: string
        type: array
    type: object
  auth.CreateSessionRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  auth.FinishSessionRequest:
    properties:
      access_token:
        type: string
    required:
    - access_token
    type: object
  auth.FinishSessionResponse:
    type: object
  auth.ResourceAccess:
    properties:
      account:
        $ref: '#/definitions/auth.Account'
    type: object
  auth.SessionInfo:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      refresh_expires_in:
        type: integer
      refresh_token:
        type: string
      scope:
        type: string
      session_state:
        type: string
      token_type:
        type: string
    type: object
  auth.UpdateSessionRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  auth.UserInfo:
    properties:
      email:
        type: string
      email_verified:
        type: boolean
      family_name:
        type: string
      given_name:
        type: string
      name:
        type: string
      preferred_username:
        type: string
      resource_access:
        $ref: '#/definitions/auth.ResourceAccess'
      username:
        type: string
    type: object
  healthcheck.Status:
    properties:
      online_time:
        type: string
    type: object
info:
  contact: {}
  description: This is a sample API for Go using Swagger
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Go Sample API
  version: "1.0"
paths:
  /:
    get:
      description: Get the status of the API
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/healthcheck.Status'
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get API status
      tags:
      - healthcheck
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return tokens
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.CreateSessionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.SessionInfo'
        "400":
          description: Invalid request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: User login
      tags:
      - authentication
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Revoke user tokens and end session
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.FinishSessionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success response
          schema:
            $ref: '#/definitions/auth.FinishSessionResponse'
        "400":
          description: Invalid request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Logout and revoke user tokens
      tags:
      - authentication
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Generate new access token using refresh token
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.UpdateSessionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.SessionInfo'
        "400":
          description: Invalid request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Refresh tokens
      tags:
      - authentication
  /auth/user:
    get:
      description: Returns information about the currently authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: User information
          schema:
            $ref: '#/definitions/auth.UserInfo'
        "401":
          description: Unauthorized - Missing or invalid token
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get user info
      tags:
      - authentication
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the access token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"